Use the author deletion authorization rule defined in an Ember-can ability class

+++ app/controllers/author/detail.js
@@ -4,6 +4,7 @@ import { inject as service } from '@ember/service';
 
 export default class AuthorDetailController extends Controller {
   @service router;
+  @service abilities;
 
   @action
   deleteAuthor(author) {
@@ -11,9 +12,11 @@ export default class AuthorDetailController extends Controller {
     // destroyRecord() and delete a record and save it right away.
     // Similar to save(), destroyRecord() also returns a promise. So, we
     // can wait for it to complete first before doing next step.
-    author.destroyRecord().then(() => {
-      // Wait for the saving to be completed first before transitioning.
-      this.router.transitionTo('author.index');
-    });
+    if (this.abilities.can('delete author', author)) {
+      author.destroyRecord().then(() => {
+        // Wait for the saving to be completed first before transitioning.
+        this.router.transitionTo('author.index');
+      });
+    }
   }
 }
